+++ src/utils/mc-utils/index.ts
@@ -1,6 +1,6 @@
 import {pinyin} from 'pinyin-pro'
 import {formatI18nKey} from '@/enum/vue-i18n-tool'
-import {textConvertAdvanced, TextConvertMode, textConvertMultipleLine} from './text-convert'
+import {copy, textConvertAdvanced, TextConvertMode, textConvertMultipleLine} from './text-convert'
 import {
   handleExportFile,
   handleImportJson,
@@ -11,6 +11,7 @@ import {
 import {showInputPrompt} from '@/components/CanUI/functions/input-prompt'
 
 export const mcUtils = {
+  copy,
   pinyin,
   formatI18nKey,
   textConvertAdvanced,